requirements.txt
rdkit>=2023.03.1
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
biopython>=1.79
meeko>=0.4.0
pyyaml>=6.0
tqdm>=4.62.0
# PyTorch packages will be installed separately in the Dockerfile
# torch>=1.10.0
# torchvision>=0.11.0
# pytorch-geometric>=2.0.0
flask>=2.0.0
flask-restful>=0.3.9
gunicorn>=20.1.0
pytest>=6.2.5
redis>=4.3.4
celery>=5.2.7
flask-cors>=3.0.10
requests>=2.27.1
joblib>=1.1.0